Extracting .../config/util/ConfigTest
